add /groups endpoint to sdk
